Card Views / Model Data Tables
After the switch from Unit Types to Unit Types with Unit Stat
associations the card view needed to be fixed. This also brought to
light the fact that a Containable find won't grab the HasMany
relationships of a 1st level BelongsTo relationship so after much head
banging against the desk the finderQuery capability was found and life
was good. Also created a bit of code that will lay out model data into a
nice little table. Team building here we come.