Team Management Functional
There's a long way for it to go in terms of usability and visuals and
even some features. But it's complete in the sense that you can add and
remove your units and choose where they start in a game. So that's sexy.
Next step, creating a game that sets units up in their starting position
and getting units moving around the board again and committing those
moves to the server.