Movement Commit Work Continues
After a move in a game, the turn of the game and the game's units is
updated along with the moved unit's position. Not enough yet though, we
need to only update the move priority and then when we update the game
turns we need to do more than change the turn variable, we need to
create a new instance of that unit with the new turn variable and the
new position to preserve replays.