Getting ready for knockout
Current plan is to stick all of the view models in the window.ViewModel
namespace so that they can be loaded using the CakePHP model name.
Convention over configuration ideally. Also the word "fuck", because I'm
not supposed to use it professionally and sometimes it's important to
remind myself that when I get home this isn't work, it's my fucking
hobby :)